+++ cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Mops CLI Changelog
 
+## Unreleased
+- Removed `mops transfer-ownership` command
+- Added `mops owner` command to manage package owners ([docs](https://docs.mops.one/cli/mops-owner))
+- Added `mops maintainers` command to manage package maintainers ([docs](https://docs.mops.one/cli/mops-maintainers))
+- Fixed bug where `mops watch` would fail if dfx.json did not exist
+
 ## 1.1.2
 - Fixed `{MOPS_ENV}` substitution in local package path